Move LookupClientAudit to its own file
